+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+Version 0.10.1
+==============
+Released: 2019-02-03
+
+ * Updated new version check to use Flathub instead of PyPI
+ * Workaround for memory leak caused by python-xlib
+   (https://github.com/python-xlib/python-xlib/issues/136)
+
 Version 0.9.0
 =============
 Released: 2019-02-02